Moved mobility/ to gaits/ & modified README.md.
Also updated the Manual Wake Up package.
