+++ data/servos_parameters.yaml
@@ -12,5 +12,5 @@ smov_states:
     ARCD: [0,   315,    220,     -1,        1] 
     ARBG: [14,  315,    440,     1,         0]
     ARBD: [1,   335,    440,     -1,        0]
-    ARJG: [13,  315,    440,     -1,        1]
-    ARJD: [2,   325 ,   390,     1,         1]
+    ARJG: [13,  315,    400,     -1,        1]
+    ARJD: [2,   315 ,   340,     1,         1]
