+++ gaits/README.md
@@ -1,0 +1,5 @@
+> The Gaits sub-directory
+
+## Overview
+
+The gaits sub-directory contains all packages and source code related to the movement of the SMOV robot in real time, with physical interactions.
