+++ gaits/manual_wake_up/package.xml
