+++ gaits/manual_wake_up/src/robot_main.cc
