+++ gaits/trigonometry/README.md
