+++ gaits/trigonometry/include/trigonometry/robot_trigonometry.h
