+++ gaits/trigonometry/package.xml
