+++ gaits/trigonometry/src/robot_trigonometry.cc
