+++ management/configuration/src/smov_configuration.cc
