+++ management/configuration/src/smov_configuration_node.cc
