--- mobility/README.md
@@ -1,8 +1,0 @@
-> The Mobility sub-directory
-
-## Overview
-
-The mobility sub-directory contains all packages and source code related to the movement of the SMOV robot in real time, with physical interactions.
-
-This is achieved through This is achieved through a package:: 
-* `The Configuration package` : depends on the Board package. It facilitates independent configuration of each servo.
